Extract URL management code from main file
